--- a/client/src/components/account.component.js
+++ b/client/src/components/account.component.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { IconButton } from '@material-ui/core';
+import { IconButton, Tooltip } from '@material-ui/core';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faGithub, faGoogle } from '@fortawesome/free-brands-svg-icons';
 import { faCheck, faInfo, faKey, faSignOutAlt, faEyeSlash, faEye } from '@fortawesome/free-solid-svg-icons';
@@ -14,45 +14,49 @@ const Account = ({ userData }) => {
     const {email, id, thirdParty, isLoading} = userData;
     const [oldPassword, setOldPassword] = useState();
     const [newPassword, setNewPassword] = useState();
-    const [confirmPsw, setConfirmPsw] = useState();
-    const [passwordVisible, setPasswordVisible] = useState(false);
-    const [newPasswordVisible, setNewPasswordVisible] = useState(false);
-    const [cfPasswordVisible, setCfPasswordVisible] = useState(false);
+    const [confirmPassword, setConfirmPassword] = useState();
+    const [disabled, setDisabled] = useState(false);
+    const [visible, setVisibility] = useState({
+        password: false,
+        newPassword: false,
+        confirmPassword: false
+    })
 
     useEffect(() => {
         const background = document.getElementById('background');
         const modal = document.getElementById('modal');
         window.onclick = function(e){
-            if(e.target === modal || e.target === background){
+            if(e.target === background && !disabled){
                 modal.classList.remove('showModal');
                 modal.classList.add('closeModal');
                 background.classList.remove('showBackground');
                 background.classList.add('hideBackground');
             }
         }
-    }, [userData])
+    }, [disabled])
 
     const submitNewPassword = (e) => {
         e.preventDefault();
-        const btn = document.getElementById('btn-changePassword');
+        const btn = document.getElementById('change-password');
         async function submitData() {
-            btn.innerHTML = "Updating...";
-            const data = { id, email, oldPassword, newPassword, confirmPassword: confirmPsw }
-            await axios.put(`${SERVER_URL}/account/user`, data, { headers: { 'X-CSRF-TOKEN': getCSRFToken()[0], 'X-XSRF-TOKEN': getCSRFToken()[1] }, withCredentials: true })
-            .then(res => setNotification(NOTIFICATION_TYPES.SUCCESS, res.data.message))
+            btn.innerHTML = "Updating..."; btn.setAttribute("disabled", "true"); btn.classList.add("disabled"); setDisabled(true);
+            const data = { id, email, oldPassword, newPassword, confirmPassword }
+            await axios.put(`${SERVER_URL}/account/user`, data, { headers: { 'XSRF-TOKEN': getCSRFToken() }, withCredentials: true })
+            .then(res => {
+                closeModal('background', 'modal');
+                setOldPassword(''); setNewPassword(''); setConfirmPassword('');
+                setNotification(NOTIFICATION_TYPES.SUCCESS, res.data.message);
+            })
             .catch(err => setNotification(NOTIFICATION_TYPES.DANGER, err.response.data.message));
-            closeModal('background', 'modal');
-            btn.removeAttribute("disabled");
-            btn.classList.remove("disabled");
-            btn.innerHTML = "Update";
-            setOldPassword(''); setNewPassword(''); setConfirmPsw('');
+            btn.innerHTML = "Update"; btn.removeAttribute("disabled"); btn.classList.remove("disabled"); setDisabled(false);
         }
-        if(!email) setNotification(NOTIFICATION_TYPES.DANGER, "Sorry, we are not able to process your request. Please try again later.")
-        else if(!oldPassword || !newPassword || !confirmPsw){ setNotification(NOTIFICATION_TYPES.DANGER, "Please Make Sure to Fill Out All Required the Fields !"); document.getElementById(!oldPassword ? 'old-password' : !newPassword ? 'new-password' : 'confirm-password').focus(); }
-        else if(oldPassword.length < 6 || newPassword.length < 6 || confirmPsw.length < 6 || oldPassword.length > 40 || newPassword.length > 40 || confirmPsw.length > 40){ setNotification(NOTIFICATION_TYPES.DANGER, 'Please Provide a Password between 6 ~ 40 characters !'); document.getElementById(oldPassword.length < 6 || oldPassword.length > 40 ? 'old-password' : newPassword.length < 6 || newPassword.length > 40 ? 'new-password' : 'confirm-password').focus(); }
-        else if(newPassword !== confirmPsw) { setNotification(NOTIFICATION_TYPES.DANGER, 'Please Make Sure Both Passwords are Match !'); document.getElementById('confirm-password').focus(); }
-        else { btn.setAttribute("disabled", "true"); btn.classList.add("disabled"); submitData(); }
+        if(!id || !email) setNotification(NOTIFICATION_TYPES.DANGER, "Sorry, we are not able to process your request. Please try again later.")
+        else if(!oldPassword || !newPassword || !confirmPassword){ setNotification(NOTIFICATION_TYPES.DANGER, "Please Make Sure to Fill Out All Required the Fields !"); document.getElementById(!oldPassword ? 'old-password' : !newPassword ? 'new-password' : 'confirm-password').focus(); }
+        else if(oldPassword.length < 6 || newPassword.length < 6 || confirmPassword.length < 6 || oldPassword.length > 40 || newPassword.length > 40 || confirmPassword.length > 40){ setNotification(NOTIFICATION_TYPES.DANGER, 'Please Provide a Password between 6 ~ 40 characters !'); document.getElementById(oldPassword.length < 6 || oldPassword.length > 40 ? 'old-password' : newPassword.length < 6 || newPassword.length > 40 ? 'new-password' : 'confirm-password').focus(); }
+        else if(newPassword !== confirmPassword) { setNotification(NOTIFICATION_TYPES.DANGER, 'Please Make Sure Both Passwords are Match !'); document.getElementById('confirm-password').focus(); }
+        else submitData();
     }
+    const handleChange = (a, b) => setVisibility({ ...visible, [a]: !b });
 
     const closeInfo = (a) => {
         const infos = document.getElementById('infos')
@@ -60,6 +64,7 @@ const Account = ({ userData }) => {
         info.classList.add('closeModal');
         setTimeout(() => { infos.removeChild(info) }, 200);
     }
+    
     return (
         <div>
             { !email ?
@@ -69,30 +74,28 @@ const Account = ({ userData }) => {
                 <div className="shape shape-3"></div>
                 <div className="shape shape-4"></div>
             </div></div>) : null }
+
             <div id="form">
                 <div id="infos">
-                    <blockquote id="info-oauth">
-                        <span><FontAwesomeIcon icon={faInfo} style={{ fontSize: '1.5em' }} /></span>
-                        <span className="info-title">Multiple OAuth Account</span>
-                        <button className="closeBtn" onClick={() => closeInfo('info-oauth')}>&times;</button>
-                        <p className="mt-10">Connecting an account with Multiple Third Parties Feature will be available in the Future Release.</p>
-                    </blockquote>
-                    <blockquote id="info-oauth">
+                    <blockquote id="info-account-verification">
                         <span><FontAwesomeIcon icon={faInfo} style={{ fontSize: '1.5em' }} /></span>
                         <span className="info-title">Account Verification</span>
-                        <button className="closeBtn" onClick={() => closeInfo('info-oauth')}>&times;</button>
-                        <p className="mt-10">Dear Users, starting from 1st May 2021, Todo Application will Verify All Accounts by sending verification through email. We are sorry to inform that unverified accounts will no longer able to use our services. More information about policy will be posted on the next release.</p>
+                        <button className="closeBtn" onClick={() => closeInfo('info-account-verification')}>&times;</button>
+                        <p className="mt-10">Dear Users,<br />Starting from <b>1<sup>st</sup> May 2021</b>, Todo Application will Verify All Accounts by sending verification through Email in order to improve our services, security, and reliablity. We are sorry to inform that unverified accounts will no longer able to use our services.</p>
                     </blockquote>
                 </div>
+
                 <div className="form__contact">
                     <div className="get_in_touch"><h1>Account</h1></div>
                     <div className="form">
-                        <div className="contact__formControl">
-                            <div className="contact__infoField">
-                                <label htmlFor="userEmail">Email Address</label>
-                                <input title="Email" id="userEmail" type="email" className="contact__inputField" value={email} disabled={true}/>
-                                <span className="contact__onFocus"></span>
-                            </div>
+                        <div className="contact__formControl contact__infoField">
+                            <label htmlFor="userEmail mt-20">Email Address</label>
+                            <input title="Email" id="userEmail" type="email" className="contact__inputField" value={email} disabled={true}/>
+                            <Tooltip placement="top" title="Verified">
+                                <IconButton className="view-eye">
+                                    <FontAwesomeIcon icon={faCheck} />
+                                </IconButton>
+                            </Tooltip>
                         </div>
                     </div>
                     <div className="oauth-container">
@@ -135,10 +138,10 @@ const Account = ({ userData }) => {
                                 <div className="contact__formControl">
                                     <div className="contact__infoField">
                                         <label htmlFor="old-password">Old Password <span className="required">*</span></label>
-                                        <input title="Old Password" id="old-password" type={ passwordVisible ? 'text':'password' } className="contact__inputField" onChange={(event) => setOldPassword(event.target.value)} value={oldPassword} spellCheck="false" autoCapitalize="none" required autoComplete={ passwordVisible ? 'off':'current-password'} />
+                                        <input title="Old Password" id="old-password" type={ visible.password ? 'text':'password' } className="contact__inputField" onChange={(event) => setOldPassword(event.target.value)} value={oldPassword} spellCheck="false" autoCapitalize="none" required autoComplete={ visible.password ? 'off':'current-password'} />
                                         <span className="contact__onFocus"></span>
-                                        <IconButton className="view-eye" onClick={() => setPasswordVisible(!passwordVisible)}>
-                                            <FontAwesomeIcon icon={passwordVisible ? faEyeSlash : faEye} />
+                                        <IconButton className="view-eye" onClick={() => handleChange('password', visible.password)}>
+                                            <FontAwesomeIcon icon={visible.password ? faEyeSlash : faEye} />
                                         </IconButton>
                                     </div>
                                 </div>
@@ -146,26 +149,28 @@ const Account = ({ userData }) => {
                                     <div className="contact__formControl">
                                         <div className="contact__infoField">
                                             <label htmlFor="new-password">New Password <span className="required">*</span></label>
-                                            <input title="New Password" id="new-password" type={ newPasswordVisible ? 'text':'password' } className="contact__inputField" onChange={(event) => setNewPassword(event.target.value)} value={newPassword} spellCheck="false" autoCapitalize="none" required autoComplete={ newPasswordVisible ? 'off':'new-password'} />
+                                            <input title="New Password" id="new-password" type={ visible.newPassword ? 'text':'password' } className="contact__inputField" onChange={(event) => setNewPassword(event.target.value)} value={newPassword} spellCheck="false" autoCapitalize="none" required autoComplete={ visible.newPassword ? 'off':'new-password'} />
                                             <span className="contact__onFocus"></span>
-                                            <IconButton className="view-eye" onClick={() => setNewPasswordVisible(!newPasswordVisible)}>
-                                                <FontAwesomeIcon icon={newPasswordVisible ? faEyeSlash : faEye} />
+                                            <IconButton className="view-eye" onClick={() => handleChange('newPassword', visible.newPassword)} name="newPassword">
+                                                <FontAwesomeIcon icon={visible.newPassword ? faEyeSlash : faEye} />
                                             </IconButton>
                                         </div>
                                     </div>
                                     <div className="contact__formControl">
                                         <div className="contact__infoField">
                                             <label htmlFor="confirm-password">Confirm New Password <span className="required">*</span></label>
-                                            <input title="Confirm New Password" id="confirm-password" type={ cfPasswordVisible ? 'text':'password' } className="contact__inputField" onChange={(event) => setConfirmPsw(event.target.value)} value={confirmPsw} spellCheck="false" autoCapitalize="none" required autoComplete={ cfPasswordVisible ? 'off':'new-password'} />
+                                            <input title="Confirm New Password" id="confirm-password" type={ visible.confirmPassword ? 'text':'password' } className="contact__inputField" onChange={(event) => setConfirmPassword(event.target.value)} value={confirmPassword} spellCheck="false" autoCapitalize="none" required autoComplete={ visible.confirmPassword ? 'off':'new-password'} />
                                             <span className="contact__onFocus"></span>
-                                            <IconButton className="view-eye" onClick={() => setCfPasswordVisible(!cfPasswordVisible)}>
-                                                <FontAwesomeIcon icon={cfPasswordVisible ? faEyeSlash : faEye} />
+                                            <IconButton className="view-eye" onClick={() => handleChange('confirmPassword', visible.confirmPassword)} name="confirmPassword">
+                                                <FontAwesomeIcon icon={visible.confirmPassword ? faEyeSlash : faEye} />
                                             </IconButton>
                                         </div>
                                     </div>
                                 </div>
                                 <p className="isCentered"><a className="animation__underline" href="/reset-password">I forgot my password</a></p>
-                                <button type="submit" id="btn-changePassword" className="btn__outline" style={{outline: 'none'}}>Update</button>
+                                <div className="inline">
+                                    <button type="submit" id="change-password" className="btn__outline" style={{outline: 'none'}}>Update</button>
+                                </div>
                             </form>
                         </div>
                     </div>