… .gitignore
… README.md
… client/.gitignore
… client/package.json
… client/public/google14f52748d0656c83.html
… client/public/index.html
… client/src/App.min.css
… client/src/components/forgot-password.component.js
… client/src/components/home.component.js
… client/src/components/login.component.js
… client/src/components/navbar.component.js
… client/src/components/register-oauth.component.js
… client/src/components/reset-password.component.js
… client/src/components/verify-account.component.js
… client/src/img/verified-success.mp4
… client/src/libraries/setNotification.js
… client/src/libraries/validation.js
… package.json
… server/.gitignore
… server/index.js
… server/lib/callback.js
… server/lib/passport.js
… server/models/blacklisted-token.model.js
… server/models/revoke-token.model.js
… server/models/todo.model.js
… server/models/token.model.js
… server/models/users.model.js
… server/package.json
… server/routes/oauth.route.js
… server/routes/todo.route.js