**What's Changed in v0.4.7:**
1. Added `reCAPTCHA` in Forgot Password Component (#102) (5a5960bec22841697331e4fda3a7a475cfcefc08)
2. Added `Encryption` for Revoked `Token` in Database (0e0413debd630dbf5a68829921db5a979b0467c6)
3. Added `504` Gateway Time Out in `Server` (f4439dc6662a5571d5de2dbde5e2fa8f7442be50)
4. Added `Verify` Status Badge on User Settings (ee1dffc7e6173989bf93b95fd915a84571e13aa4)
5. Added `Verify` Account Feature (#100) (04116319aabd6f2bfce1836e29cc86e8ae9ab214) (17d583e5883e5f6219cf14c09e1235082c4f519b)
6. Added Google Site Verification (516c07337c5dce35579d42f75d75bcc843d11cc5)
7. Upgrade Dependency to the Latest Version (0cdc674e93534a38340c24880ad36acec2e5801c)
8. Improved Performance on Todo Router using `Passport` (#105) (55b0c341b17eb581955f3e67ab7e901203d8743f)
9. Improved Security in Server API (9f8de47ff2b3b110a36092b78807203c0310d728)
10. Optimized Server API (9f8de47ff2b3b110a36092b78807203c0310d728)
11. Optimized and Improved `Client` Component (e44d094bc5ab3ae2272f345aa01b50a80f3aac01)
12. Improved JSON Response (#103) (45ab2e9c6a92791d0460513355615b93ea605e62)
13. Improved Accessibility (6b7242f944faef1f44e9f45d22bb1f015ccf03f0)
14. Changed Method of Clearing `httpOnly` Cookie from `res.clearCookie` to `res.cookie` (0fcb80a586035fbb14eb4209d9d1c5b70a7496d3)
15. Resolved Issues of:
    * JWT Cookie Not Being Cleared After Logout (#99) (630424cc5e726155a807b755a0a0884d3428870c)
    * App Crashes when Connecting Account with GitHub (#98) (e8c1850bdaabefdf70e6d7d5a84ce822b673f6fc)
    * Severity in Forget Password API (#106) (57d242a2decd63e6cfcfd4b82152cab469a56d47)
    * Confirm Password not Being Validated on OAuth Registration (#104) (b3db8d4ddbb89b484cf6090e289388b9c3934ba8)
    * Minor Issues

Details About Changes in v0.4.7:
| Client | `build\static\js\2.6ca8d93c.chunk.js` | `build\static\js\main.2591d28a.chunk.js` | `build\static\css\main.030489ff.chunk.css` | `build\static\js\3.668acfa5.chunk.js` | `build\static\js\runtime-main.3cbd5898.js`
------ | --------------------------------------- | ------------------------------------------- | ---------------------------------------------- | -------------------------------------------------- | ---------------------------------------------- |
| v0.4.7 | 148.29 kB (-1.25 kB) | 26.54 kB (+478 B) | 4.47 kB (+72 B) | 1.56 kB | 1.16 kB |
| v0.4.6 | 147.82 kB | 26.07 kB | 4.4 kB | 1.56 kB | 1.16 kB |
| v0.4.5 | 106.38 kB | 25.61 kB | 4.29 kB | 1.56 kB | 1.16 kB |

| Server | Size |
| ------- | ---- |
| v0.4.7 | 190 KB (+14 kB |
| v0.4.6 | 176 kB |

For more information about changes in this release, please visit our [project roadmaps v0.4.6](https://github.com/stanleyowen/todo-application/projects/26)